+++ proto/tests/common/mod.rs
@@ -206,6 +206,12 @@ impl services::ServiceSink for ServiceSinkWrapper {
     async fn send(&self, d: hardy_bpa::Bytes) -> services::Result<hardy_bpv7::bundle::Id> {
         self.0.send(d).await
     }
+    async fn send_streamed(
+        &self,
+        s: &dyn hardy_bpa::stream::Receiver<hardy_bpa::stream::Segment>,
+    ) -> services::Result<hardy_bpv7::bundle::Id> {
+        self.0.send_streamed(s).await
+    }
     async fn cancel(&self, id: &hardy_bpv7::bundle::Id) -> services::Result<bool> {
         self.0.cancel(id).await
     }